requirements.txt
Flask==3.0.3
requests
python-telegram-bot==20.6
